requirements.txt
PyYAML
flask
flask_cors
flask_sslify
flask-restx
flask-swagger
flask-swagger-ui
Flask-JWT-Extended
Flask-HTTPAuth
Flask-SQLAlchemy
#Flask-MongoAlchemy
flask-mongoengine
Flask-WTF
Flask-Login
flask-marshmallow
flask-bcrypt
Flask-Mail
SQLAlchemy
marshmallow
marshmallow-sqlalchemy
mysql-connector-python
PyMySQL
datastore-entity
google-cloud-firestore
backports-datetime-fromisoformat
fireo
urllib3
requests
webargs
apispec
apispec-webframeworks
gunicorn